Change validation in CreateAdPage
Uses the disabled-attribute on the submit-button instead of alerts to
show validation errors. Also sets the hover-style on the button to
only show if it is enabled.

In the future we should also show validation errors in the form itself,
like on finn.no's page for creating ads, since the disabled button
doesnt explicitly say what is wrong with the form.