Add simple tests for StatsPage
I would like to also test the components (StatsElement), but these are dynamic,
and aren't rendered unless the currently logged in user has any ads saved.
Since there is no logged in user when the tests are ran, these don't show.
Maybe I'll find a way to test it in the future.

Relates: #66